requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
Click==7.0
cryptography==2.3.1
dnspython==1.15.0
enum34==1.1.6
eventlet==0.24.1
Flask==1.0.2
Flask-SocketIO==3.0.2
future==0.16.0
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
monotonic==1.5
pycparser==2.19
python-engineio==2.3.2
python-socketio==2.0.0
python-telegram-bot==11.1.0
six==1.11.0
Werkzeug==0.14.1
